+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 * **x.X versions** are either updates that add content or major bug fixes
 * **x.x.X versions** are either small content update (language translation, keybind change, ...) or bug fixes
 
+## v1.5.4 ([@naturecodevoid](https://github.com/naturecodevoid))
+
+* Added volume increment setting
+* Added setting that allows user to increment the music volume up to 200% (disabled by default)
+* Improved config: added categories and tooltips
+* Improved volume display by adding percent symbol to the end of the displayed text
+
 ## v1.5.3
 
 * Fixed custom category auto changing after a music was played, when not in creative or cheat disabled